Suppress the warning message on 2.6.32+ kernels
Writing to "/sys/DEVICE/power/level" on a 2.6.32 or later kernel
triggers a kernel warning message that says this interface is
deprecated. So now if "/sys/DEVICE/power/control" exists we do nothing
(there's nothing we need to do anyway).

We should still support pre 2.6.32 kernels but I don't have a machine
to test. Welcome feedbacks.